Improve session test readability and maintainability.
